feat(client): add invitations utils file
