+++ packages/client/src/pages/authorized/admin/Invitations/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { CSSProperties } from 'styled-components';
 
 import {
   Button,
@@ -9,40 +8,10 @@ import {
   InvitationStatusField,
   TextField,
 } from '../../../../elements';
-import { InvitationStatus } from '../../../../enums/invitationStatus.enum';
 import { useMediaQuery } from '../../../../hooks';
-import { ApiInvitation } from '../../../../interfaces/api.types';
-import { ITheme, ThemeType } from '../../../../theme/Theme';
+import { ThemeType } from '../../../../theme/Theme';
 import { CreateButtonContainer, ListContainer, Wrapper } from './Invitations.styled';
-
-const getRowStyle = (
-  { status, expirationDate }: ApiInvitation,
-  { palette }: ITheme,
-): CSSProperties => {
-  const baseStyling: CSSProperties = {
-    borderLeftWidth: 7,
-    borderLeftStyle: 'solid',
-  };
-
-  if (expirationDate < new Date()) {
-    return {
-      ...baseStyling,
-      borderLeftColor: palette.colors.red,
-    };
-  }
-
-  if (status === InvitationStatus.Accepted) {
-    return {
-      ...baseStyling,
-      borderLeftColor: palette.primary.light,
-    };
-  }
-
-  return {
-    ...baseStyling,
-    borderLeftColor: palette.colors.orange,
-  };
-};
+import { getRowStyle } from './Invitations.utils';
 
 const Invitations = () => {
   const isMobile = useMediaQuery(({ breakpoints, down }) => down(breakpoints.lg));
